added email, password and table migration and dtos for user object

+++ src/main/kotlin/com/coded/spring/ordering/domain/projections/MenuBasicInfoProjection.kt
@@ -1,7 +1,9 @@
 package com.coded.spring.ordering.domain.projections
 
+import java.math.BigDecimal
+
 interface MenuBasicInfoProjection {
     val name: String
     val id: Long
-    val price: Double
+    val price: BigDecimal
 }
