+++ src/main/kotlin/com/coded/spring/ordering/domain/requests/MenuItemCreateRequestDto.kt
@@ -2,11 +2,12 @@ package com.coded.spring.ordering.domain.requests
 
 import com.coded.spring.ordering.domain.entities.Menu
 import com.coded.spring.ordering.domain.entities.Restaurant
+import java.math.BigDecimal
 
 data class MenuCreateRequestDto(
     val name: String,
     val restaurantId: Long,
-    val price: Double
+    val price: BigDecimal,
 )
 
 fun MenuCreateRequestDto.toEntity(restaurant: Restaurant): Menu = Menu(
